pyproject.toml
[project]
name = "rejeki"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["fastmcp", "python-dotenv"]

[project.scripts]
rejeki = "rejeki.server:main"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
